Add files via upload
*Problem Lab Week 3*

Some students are standing in line for PT (Descending Order). If the student in front of another student is 20% taller, the latter student will not face the sun.
The heights of the students will be taken as input. You need to check whether a student will face the sun or not.